requirements.txt
numpy==1.24.3
Pillow==10.0.0
chromadb==0.4.22
requests==2.31.0
PyMuPDF==1.23.8
sentence-transformers
onnxruntime==1.15.1
torch==2.0.1
transformers==4.33.2